Issue #14769: Implemented sizes package structure for IT Regression area
