--- src/it/resources/org/checkstyle/suppressionxpathfilter/executablestatementcount/InputXpathExecutableStatementCountCustomMax.java
@@ -1,14 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.executablestatementcount;
-
-public class InputXpathExecutableStatementCountCustomMax {
-    public InputXpathExecutableStatementCountCustomMax() // warn
-    {
-        int i = 1;
-        if (System.currentTimeMillis() == 0) {
-        } else if (System.currentTimeMillis() == 0) {
-        } else {
-        }
-    }
-    /** Empty constructor */
-    public InputXpathExecutableStatementCountCustomMax(int i) {}
-}
