--- src/it/resources/org/checkstyle/suppressionxpathfilter/executablestatementcount/InputXpathExecutableStatementCountDefault.java
@@ -1,17 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.executablestatementcount;
-
-public class InputXpathExecutableStatementCountDefault {
-    public void ElseIfLadder() { // warn
-        if (System.currentTimeMillis() == 0) {
-        } else {
-            if (System.currentTimeMillis() == 0) {
-            } else {
-                if (System.currentTimeMillis() == 0) {
-                }
-            }
-
-            if (System.currentTimeMillis() == 0) {
-            }
-        }
-    }
-}
