--- src/it/resources/org/checkstyle/suppressionxpathfilter/lambdabodylength/InputXpathLambdaBodyLengthCustomMax.java
@@ -1,16 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.lambdabodylength;
-
-import java.util.function.Function;
-
-public class InputXpathLambdaBodyLengthCustomMax {
-    void test() {
-        Runnable r = () -> { // warn
-            method(2);
-            method(3);
-            method(4);
-            method(5);
-        };
-    }
-
-    void method(int i) {}
-}
