--- src/it/resources/org/checkstyle/suppressionxpathfilter/methodcount/InputXpathMethodCountDefault.java
@@ -1,6 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.methodcount;
-
-class InputXpathMethodCountDefault { // warn
-    void foo() {}
-    void foo1() {}
-}
