--- src/it/resources/org/checkstyle/suppressionxpathfilter/methodcount/InputXpathMethodCountPrivate.java
@@ -1,6 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.methodcount;
-
-class InputXpathMethodCountPrivate { // warn
-    private void foo() {}
-    private void foo1() {}
-}
