--- src/it/resources/org/checkstyle/suppressionxpathfilter/methodcount/InputXpathMethodCountProtected.java
@@ -1,6 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.methodcount;
-
-class InputXpathMethodCountProtected { // warn
-    protected void foo() {}
-    protected void foo1() {}
-}
