--- src/it/resources/org/checkstyle/suppressionxpathfilter/methodcount/InputXpathMethodCountPublic.java
@@ -1,6 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.methodcount;
-
-class InputXpathMethodCountPublic { // warn
-    public void foo() {}
-    public void foo1() {}
-}
