--- src/it/resources/org/checkstyle/suppressionxpathfilter/methodlength/InputXpathMethodLengthNoEmptyLines.java
@@ -1,22 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.methodlength;
-
-public class InputXpathMethodLengthNoEmptyLines {
-
-    public InputXpathMethodLengthNoEmptyLines(){
-        // a line
-        // a line
-        // a line
-        // a line
-        // a line
-
-
-    }
-
-    protected void methodOne(){ // warn
-        if(true){
-            if(true){
-            }
-        }
-    }
-
-}
