--- src/it/resources/org/checkstyle/suppressionxpathfilter/methodlength/InputXpathMethodLengthSimple.java
@@ -1,27 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.methodlength;
-
-public class InputXpathMethodLengthSimple {
-    protected InputXpathMethodLengthSimple() { // warn
-        // a line
-        // a line
-        // a line
-        // a line
-        // a line
-        // a line
-        // a line
-        // a line
-
-    }
-
-    private void methodOne(){
-        // a line
-        // a line
-        // a line
-        // a line
-        // a line
-        // a line
-        // a line
-        // a line
-    }
-
-}
