--- src/it/resources/org/checkstyle/suppressionxpathfilter/methodlength/InputXpathMethodLengthSingleToken.java
@@ -1,13 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.methodlength;
-
-public class InputXpathMethodLengthSingleToken {
-
-    public void methodOne() { }
-
-    InputXpathMethodLengthSingleToken a =
-            new InputXpathMethodLengthSingleToken(){
-        public void methodOne(){ // warn
-
-        }
-    };
-}
