--- src/it/resources/org/checkstyle/suppressionxpathfilter/outertypenumber/InputXpathOuterTypeNumberDefault.java
@@ -1,9 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.outertypenumber; // warn
-
-public class InputXpathOuterTypeNumberDefault {
-    public void test() {}
-}
-
-interface Constants {
-    int X = 10;
-}
