--- src/it/resources/org/checkstyle/suppressionxpathfilter/parameternumber/InputXpathParameterNumberIgnoreOverriddenMethods.java
@@ -1,14 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.parameternumber;
-
-public class InputXpathParameterNumberIgnoreOverriddenMethods
-        extends InputXpathParameterNumberDefault {
-
-     public InputXpathParameterNumberIgnoreOverriddenMethods(int a, // warn
-                    int b, int c, int d, int e, int f, int g, int h)
-    {
-    }
-    @Override
-    void myMethod(int a, int b, int c, int d, int e, int f, int g, int h, // ok
-                    int k, int l, int m) {
-    }
-}
