--- src/it/resources/org/checkstyle/suppressionxpathfilter/recordcomponentnumber/InputXpathRecordComponentNumberCustomMax.java
@@ -1,10 +1,0 @@
-// Java17
-package org.checkstyle.suppressionxpathfilter.recordcomponentnumber;
-
-/* Config:
- *
- * max = 1
- */
-public class InputXpathRecordComponentNumberCustomMax {
-    public record MyRecord(int x, int y) { } // warn
-}
