--- src/it/resources/org/checkstyle/suppressionxpathfilter/recordcomponentnumber/InputXpathRecordComponentNumberDefault.java
@@ -1,15 +1,0 @@
-// Java17
-package org.checkstyle.suppressionxpathfilter.recordcomponentnumber;
-
-/* Config:
- *
- * max = 8
- */
-public record InputXpathRecordComponentNumberDefault(int x, int y, int z, // warn
-                                                int a, int b, int c,
-                                                int d, int e, int f,
-                                                int g, int h, int i,
-                                                int j, int k,
-                                                 String... myVarargs) {
-
-}
