+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sizes/anoninnerlength/InputXpathAnonInnerLength.java
@@ -1,0 +1,15 @@
+package org.checkstyle.suppressionxpathfilter.sizes.anoninnerlength;
+
+import java.util.Comparator;
+
+public class InputXpathAnonInnerLength {
+    public int compare(String v1, String v2) {
+        Comparator<String> comp = new Comparator<String>() { // warn: inner class is 6 lines (max=5)
+            @Override
+            public int compare(String s1, String s2) {
+                return s1.compareTo(s2);
+            }
+        };
+        return comp.compare(v1, v2);
+    }
+}
