+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sizes/executablestatementcount/InputXpathExecutableStatementCountCustomMax.java
@@ -1,0 +1,14 @@
+package org.checkstyle.suppressionxpathfilter.sizes.executablestatementcount;
+
+public class InputXpathExecutableStatementCountCustomMax {
+    public InputXpathExecutableStatementCountCustomMax() // warn
+    {
+        int i = 1;
+        if (System.currentTimeMillis() == 0) {
+        } else if (System.currentTimeMillis() == 0) {
+        } else {
+        }
+    }
+    /** Empty constructor */
+    public InputXpathExecutableStatementCountCustomMax(int i) {}
+}
