+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sizes/executablestatementcount/InputXpathExecutableStatementCountDefault.java
@@ -1,0 +1,17 @@
+package org.checkstyle.suppressionxpathfilter.sizes.executablestatementcount;
+
+public class InputXpathExecutableStatementCountDefault {
+    public void ElseIfLadder() { // warn
+        if (System.currentTimeMillis() == 0) {
+        } else {
+            if (System.currentTimeMillis() == 0) {
+            } else {
+                if (System.currentTimeMillis() == 0) {
+                }
+            }
+
+            if (System.currentTimeMillis() == 0) {
+            }
+        }
+    }
+}
