+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sizes/lambdabodylength/InputXpathLambdaBodyLengthCustomMax.java
@@ -1,0 +1,16 @@
+package org.checkstyle.suppressionxpathfilter.sizes.lambdabodylength;
+
+import java.util.function.Function;
+
+public class InputXpathLambdaBodyLengthCustomMax {
+    void test() {
+        Runnable r = () -> { // warn
+            method(2);
+            method(3);
+            method(4);
+            method(5);
+        };
+    }
+
+    void method(int i) {}
+}
