+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sizes/lambdabodylength/InputXpathLambdaBodyLengthDefaultMax.java
@@ -1,0 +1,20 @@
+package org.checkstyle.suppressionxpathfilter.sizes.lambdabodylength;
+
+import java.util.function.Function;
+
+public class InputXpathLambdaBodyLengthDefaultMax {
+    void test() {
+        Function<String, String> trimmer = (s) -> // warn
+            s.trim()
+             .trim()
+             .trim()
+             .trim()
+             .trim()
+             .trim()
+             .trim()
+             .trim()
+             .trim()
+             .trim()
+             .trim();
+    }
+}
