+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sizes/methodcount/InputXpathMethodCountDefault.java
@@ -1,0 +1,6 @@
+package org.checkstyle.suppressionxpathfilter.sizes.methodcount;
+
+class InputXpathMethodCountDefault { // warn
+    void foo() {}
+    void foo1() {}
+}
