+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sizes/methodcount/InputXpathMethodCountPrivate.java
@@ -1,0 +1,6 @@
+package org.checkstyle.suppressionxpathfilter.sizes.methodcount;
+
+class InputXpathMethodCountPrivate { // warn
+    private void foo() {}
+    private void foo1() {}
+}
