+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sizes/methodcount/InputXpathMethodCountProtected.java
@@ -1,0 +1,6 @@
+package org.checkstyle.suppressionxpathfilter.sizes.methodcount;
+
+class InputXpathMethodCountProtected { // warn
+    protected void foo() {}
+    protected void foo1() {}
+}
