+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sizes/methodcount/InputXpathMethodCountPublic.java
@@ -1,0 +1,6 @@
+package org.checkstyle.suppressionxpathfilter.sizes.methodcount;
+
+class InputXpathMethodCountPublic { // warn
+    public void foo() {}
+    public void foo1() {}
+}
