+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sizes/methodlength/InputXpathMethodLengthSimple.java
@@ -1,0 +1,27 @@
+package org.checkstyle.suppressionxpathfilter.sizes.methodlength;
+
+public class InputXpathMethodLengthSimple {
+    protected InputXpathMethodLengthSimple() { // warn
+        // a line
+        // a line
+        // a line
+        // a line
+        // a line
+        // a line
+        // a line
+        // a line
+
+    }
+
+    private void methodOne(){
+        // a line
+        // a line
+        // a line
+        // a line
+        // a line
+        // a line
+        // a line
+        // a line
+    }
+
+}
