+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sizes/methodlength/InputXpathMethodLengthSingleToken.java
@@ -1,0 +1,13 @@
+package org.checkstyle.suppressionxpathfilter.sizes.methodlength;
+
+public class InputXpathMethodLengthSingleToken {
+
+    public void methodOne() { }
+
+    InputXpathMethodLengthSingleToken a =
+            new InputXpathMethodLengthSingleToken(){
+        public void methodOne(){ // warn
+
+        }
+    };
+}
