+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sizes/outertypenumber/InputXpathOuterTypeNumber.java
@@ -1,0 +1,30 @@
+package org.checkstyle.suppressionxpathfilter.sizes.outertypenumber; //warn
+
+public class InputXpathOuterTypeNumber {
+    int i; int j;
+}
+class InputOuterTypeNumberSimple2
+{
+    /** Some more Javadoc. */
+    public void doSomething()
+    {
+        //"O" should be named "o"
+        for (Object O : new java.util.ArrayList())
+        {
+
+        }
+    }
+}
+
+/** Test enum for member naming check */
+enum MyEnum1
+{
+    /** ABC constant */
+    ABC,
+
+    /** XYZ constant */
+    XYZ;
+
+    /** Should be mSomeMember */
+    private int someMember;
+}
