+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sizes/outertypenumber/InputXpathOuterTypeNumberDefault.java
@@ -1,0 +1,9 @@
+package org.checkstyle.suppressionxpathfilter.sizes.outertypenumber; // warn
+
+public class InputXpathOuterTypeNumberDefault {
+    public void test() {}
+}
+
+interface Constants {
+    int X = 10;
+}
