+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sizes/parameternumber/InputXpathParameterNumberIgnoreOverriddenMethods.java
@@ -1,0 +1,14 @@
+package org.checkstyle.suppressionxpathfilter.sizes.parameternumber;
+
+public class InputXpathParameterNumberIgnoreOverriddenMethods
+        extends InputXpathParameterNumberDefault {
+
+     public InputXpathParameterNumberIgnoreOverriddenMethods(int a, // warn
+                    int b, int c, int d, int e, int f, int g, int h)
+    {
+    }
+    @Override
+    void myMethod(int a, int b, int c, int d, int e, int f, int g, int h, // ok
+                    int k, int l, int m) {
+    }
+}
