+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sizes/recordcomponentnumber/InputXpathRecordComponentNumberCustomMax.java
@@ -1,0 +1,10 @@
+// Java17
+package org.checkstyle.suppressionxpathfilter.sizes.recordcomponentnumber;
+
+/* Config:
+ *
+ * max = 1
+ */
+public class InputXpathRecordComponentNumberCustomMax {
+    public record MyRecord(int x, int y) { } // warn
+}
