+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sizes/recordcomponentnumber/InputXpathRecordComponentNumberDefault.java
@@ -1,0 +1,15 @@
+// Java17
+package org.checkstyle.suppressionxpathfilter.sizes.recordcomponentnumber;
+
+/* Config:
+ *
+ * max = 8
+ */
+public record InputXpathRecordComponentNumberDefault(int x, int y, int z, // warn
+                                                int a, int b, int c,
+                                                int d, int e, int f,
+                                                int g, int h, int i,
+                                                int j, int k,
+                                                 String... myVarargs) {
+
+}
